Closes #8
* consolidate @models/user and @states/user
* remove @states
* refactored tests
