+++ public/index.html
@@ -55,8 +55,6 @@
 
           "@models/user": "/ts-react-playground/script/models/user.js",
 
-          "@states/user": "/ts-react-playground/script/states/user.js",
-
           "@components/layout": "/ts-react-playground/script/components/layout/index.js",
 
           "@atoms/button": "/ts-react-playground/script/components/atoms/button/index.js",
