+++ src/script/components/atoms/button/index.spec.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { create } from 'react-test-renderer';
 import { mount } from 'enzyme';
 
-import * as TestSubject from './index';
+import * as TestSubject from '@atoms/button';
 
 describe(`${TestSubject.Button.displayName}`, () => {
   it('exports default', () => {
