+++ src/script/components/organisms/main/index.spec.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { create } from 'react-test-renderer';
 import { mount } from 'enzyme';
 
-import * as TestSubject from './index';
+import * as TestSubject from '@organisms/main';
 
 describe(`${TestSubject.Main.displayName}`, () => {
   it('exports default', () => {
