+++ src/script/models/user.tsx
@@ -1,3 +1,4 @@
+import { atom, selector } from 'recoil';
 import { number as IsNumber, shape as IsShape, string as IsString } from 'prop-types';
 import faker from 'faker';
 
@@ -48,4 +49,21 @@ export const generateAsync = async (count = 100): Promise<User[]> => {
   return generate(count);
 };
 
-export default User;
+const USERS_KEY = 'userAtom';
+
+export const UsersAtom = atom({
+  key: USERS_KEY,
+  default: [] as User[],
+});
+
+const USERS_LENGTH_KEY = 'usersLengthSelector';
+
+export const UsersLength = selector({
+  key: USERS_LENGTH_KEY,
+  get: ({ get }) => {
+    const users = get(UsersAtom);
+    return users.length;
+  },
+});
+
+export default UsersAtom;
